bugfix: line breaks on windows in bash script
on Charlie's Windows computer the following error message occurred for each empty line:
update-build-info.sh: line *: $'\r': command not found

The problem was that git checks out \n as \r\n because it is on Windows but bash (as it is a Linux program) expects only \n. So it reads lines which are suppossed to be empty as containing a command consisting of one character: \r.
Similarly, the arguments on the non-empty lines are wrong.

These changes shall tell git to not convert \n in shell scripts.

copied from:
https://techblog.dorogin.com/case-of-windows-line-ending-in-bash-script-7236f056abe